change: Private / Public router 분리해서 관리
