+++ src/router/PrivateRouter.jsx
@@ -1,0 +1,11 @@
+import { LOCAL_STORAGE } from 'constants';
+import { ROUTES, useLocalStorage } from 'hooks';
+import { Navigate, Outlet } from 'react-router-dom';
+
+const PrivateRouter = () => {
+  const { storageValue: accessToken } = useLocalStorage(LOCAL_STORAGE.ACCESS_TOKEN);
+
+  return !accessToken ? <Navigate to={ROUTES.HOME} replace /> : <Outlet />;
+};
+
+export default PrivateRouter;
