+++ src/router/Router.jsx
@@ -1,25 +1,10 @@
-import { LOCAL_STORAGE, ROUTES } from 'constants';
-import { useLocalStorage } from 'hooks';
+import { ROUTES } from 'constants';
 import AccountPage from 'pages/AccountPage';
 import TodoPage from 'pages/ToDoPage';
-import { Navigate, Outlet, createBrowserRouter, useLocation } from 'react-router-dom';
+import { createBrowserRouter } from 'react-router-dom';
 
-const PrivateRouter = () => {
-  const { storageValue: accessToken } = useLocalStorage(LOCAL_STORAGE.ACCESS_TOKEN);
-
-  return !accessToken ? <Navigate to={ROUTES.HOME} replace /> : <Outlet />;
-};
-
-const PublicRouter = () => {
-  const { pathname } = useLocation();
-  const { storageValue: accessToken } = useLocalStorage(LOCAL_STORAGE.ACCESS_TOKEN);
-
-  if (pathname === ROUTES.HOME) {
-    return accessToken ? <Navigate to={ROUTES.TODO} replace /> : <Outlet />;
-  }
-
-  return accessToken ? <Navigate to={ROUTES.HOME} replace /> : <Outlet />;
-};
+import PrivateRouter from './PrivateRouter';
+import PublicRouter from './PublicRouter';
 
 export const router = createBrowserRouter([
   {
